MAINT create BinaryClassifierCurveDisplayMixin
