+++ sklearn/metrics/_plot/det_curve.py
@@ -1,13 +1,10 @@
 import scipy as sp
 
 from .. import det_curve
-from .._base import _check_pos_label_consistency
-
-from ...utils import check_matplotlib_support
-from ...utils._response import _get_response_values_binary
+from ...utils._plot import BinaryClassifierCurveDisplayMixin
 
 
-class DetCurveDisplay:
+class DetCurveDisplay(BinaryClassifierCurveDisplayMixin):
     """DET curve visualization.
 
     It is recommend to use :func:`~sklearn.metrics.DetCurveDisplay.from_estimator`
@@ -163,15 +160,13 @@ class DetCurveDisplay:
         <...>
         >>> plt.show()
         """
-        check_matplotlib_support(f"{cls.__name__}.from_estimator")
-
-        name = estimator.__class__.__name__ if name is None else name
-
-        y_pred, pos_label = _get_response_values_binary(
+        y_pred, pos_label, name = super().from_estimator(
             estimator,
             X,
-            response_method,
+            y,
+            response_method=response_method,
             pos_label=pos_label,
+            name=name,
         )
 
         return cls.from_predictions(
@@ -259,7 +254,10 @@ class DetCurveDisplay:
         <...>
         >>> plt.show()
         """
-        check_matplotlib_support(f"{cls.__name__}.from_predictions")
+        pos_label_validated, name = super().from_predictions(
+            y_true, y_pred, sample_weight=sample_weight, pos_label=pos_label, name=name
+        )
+
         fpr, fnr, _ = det_curve(
             y_true,
             y_pred,
@@ -267,14 +265,11 @@ class DetCurveDisplay:
             sample_weight=sample_weight,
         )
 
-        pos_label = _check_pos_label_consistency(pos_label, y_true)
-        name = "Classifier" if name is None else name
-
         viz = DetCurveDisplay(
             fpr=fpr,
             fnr=fnr,
             estimator_name=name,
-            pos_label=pos_label,
+            pos_label=pos_label_validated,
         )
 
         return viz.plot(ax=ax, name=name, **kwargs)
@@ -300,9 +295,8 @@ class DetCurveDisplay:
         display : :class:`~sklearn.metrics.plot.DetCurveDisplay`
             Object that stores computed values.
         """
-        check_matplotlib_support("DetCurveDisplay.plot")
+        name = super().plot(name=name)
 
-        name = self.estimator_name if name is None else name
         line_kwargs = {} if name is None else {"label": name}
         line_kwargs.update(**kwargs)
